Move FileExtensions to logic.util to prevent architecture test from failing Plain text with uppercase P
Importer now use FileExtensions enum, too
Grouped Importer in FileExtension
Created test for FileExtions
