+++ src/main/java/net/sf/jabref/gui/preftabs/PreferencesDialog.java
@@ -37,8 +37,6 @@ import javax.swing.ListSelectionModel;
 
 import net.sf.jabref.Globals;
 import net.sf.jabref.JabRefException;
-
-import net.sf.jabref.gui.FileExtensions;
 import net.sf.jabref.gui.GUIGlobals;
 import net.sf.jabref.gui.JabRefFrame;
 import net.sf.jabref.gui.NewFileDialogs;
@@ -46,6 +44,7 @@ import net.sf.jabref.gui.keyboard.KeyBinder;
 import net.sf.jabref.gui.maintable.MainTable;
 import net.sf.jabref.logic.exporter.ExportFormats;
 import net.sf.jabref.logic.l10n.Localization;
+import net.sf.jabref.logic.util.FileExtensions;
 import net.sf.jabref.preferences.JabRefPreferences;
 import net.sf.jabref.preferences.JabRefPreferencesFilter;
