+++ src/main/java/net/sf/jabref/gui/protectedterms/NewProtectedTermsFileDialog.java
@@ -13,14 +13,13 @@ import javax.swing.JDialog;
 import javax.swing.JTextField;
 
 import net.sf.jabref.Globals;
-
-import net.sf.jabref.gui.FileExtensions;
 import net.sf.jabref.gui.JabRefFrame;
 import net.sf.jabref.gui.actions.BrowseAction;
 import net.sf.jabref.gui.keyboard.KeyBinding;
 
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.logic.protectedterms.ProtectedTermsLoader;
+import net.sf.jabref.logic.util.FileExtensions;
 
 import com.jgoodies.forms.builder.ButtonBarBuilder;
 import com.jgoodies.forms.builder.FormBuilder;
