+++ src/main/java/net/sf/jabref/logic/importer/fileformat/BibTeXMLImporter.java
@@ -25,9 +25,9 @@ import java.util.regex.Pattern;
 import javax.xml.parsers.SAXParser;
 import javax.xml.parsers.SAXParserFactory;
 
-import net.sf.jabref.gui.FileExtensions;
 import net.sf.jabref.logic.importer.ParserResult;
 import net.sf.jabref.logic.importer.util.BibTeXMLHandler;
+import net.sf.jabref.logic.util.FileExtensions;
 import net.sf.jabref.model.entry.BibEntry;
 
 import org.apache.commons.logging.Log;
