+++ src/main/java/net/sf/jabref/logic/importer/fileformat/CopacImporter.java
@@ -22,8 +22,8 @@ import java.util.List;
 import java.util.Objects;
 import java.util.regex.Pattern;
 
-import net.sf.jabref.gui.FileExtensions;
 import net.sf.jabref.logic.importer.ParserResult;
+import net.sf.jabref.logic.util.FileExtensions;
 import net.sf.jabref.model.entry.BibEntry;
 import net.sf.jabref.model.entry.FieldName;
