+++ src/main/java/net/sf/jabref/logic/importer/fileformat/MedlineImporter.java
@@ -33,7 +33,6 @@ import javax.xml.stream.XMLInputFactory;
 import javax.xml.stream.XMLStreamException;
 import javax.xml.stream.XMLStreamReader;
 
-import net.sf.jabref.gui.FileExtensions;
 import net.sf.jabref.logic.importer.ParserResult;
 import net.sf.jabref.logic.importer.fileformat.medline.Abstract;
 import net.sf.jabref.logic.importer.fileformat.medline.AbstractText;
@@ -79,6 +78,7 @@ import net.sf.jabref.logic.importer.fileformat.medline.QualifierName;
 import net.sf.jabref.logic.importer.fileformat.medline.Section;
 import net.sf.jabref.logic.importer.fileformat.medline.Sections;
 import net.sf.jabref.logic.importer.fileformat.medline.Text;
+import net.sf.jabref.logic.util.FileExtensions;
 import net.sf.jabref.logic.util.strings.StringUtil;
 import net.sf.jabref.model.entry.BibEntry;
 import net.sf.jabref.model.entry.FieldName;
