+++ src/main/java/net/sf/jabref/logic/importer/fileformat/RepecNepImporter.java
@@ -31,9 +31,9 @@ import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Objects;
 
-import net.sf.jabref.gui.FileExtensions;
 import net.sf.jabref.logic.importer.ImportFormatPreferences;
 import net.sf.jabref.logic.importer.ParserResult;
+import net.sf.jabref.logic.util.FileExtensions;
 import net.sf.jabref.model.entry.BibEntry;
 import net.sf.jabref.model.entry.FieldName;
 import net.sf.jabref.model.entry.IdGenerator;
