+++ src/test/java/net/sf/jabref/logic/importer/fileformat/FreeCiteImporterTest.java
@@ -1,8 +1,7 @@
 package net.sf.jabref.logic.importer.fileformat;
 
-import java.util.Arrays;
-
 import net.sf.jabref.logic.importer.ImportFormatPreferences;
+import net.sf.jabref.logic.util.FileExtensions;
 import net.sf.jabref.preferences.JabRefPreferences;
 
 import org.junit.Before;
@@ -13,6 +12,7 @@ import static org.junit.Assert.assertEquals;
 public class FreeCiteImporterTest {
 
     private FreeCiteImporter importer;
+
 
     @Before
     public void setUp() {
@@ -26,11 +26,12 @@ public class FreeCiteImporterTest {
 
     @Test
     public void testsGetExtensions() {
-        assertEquals(Arrays.asList(".txt",".xml"), importer.getExtensions());
+        assertEquals(FileExtensions.FREECITE, importer.getExtensions());
     }
 
     @Test
     public void testGetDescription() {
-        assertEquals("This importer parses text format citations using the online API of FreeCite.", importer.getDescription());
+        assertEquals("This importer parses text format citations using the online API of FreeCite.",
+                importer.getDescription());
     }
 }
