+++ src/test/java/net/sf/jabref/logic/importer/fileformat/MsBibImporterTest.java
@@ -10,7 +10,7 @@ import java.util.Collections;
 import java.util.List;
 
 import net.sf.jabref.Globals;
-import net.sf.jabref.gui.FileExtensions;
+import net.sf.jabref.logic.util.FileExtensions;
 import net.sf.jabref.model.entry.BibEntry;
 import net.sf.jabref.preferences.JabRefPreferences;
